# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口文档",
        "title": "AB迎新系统接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "123.207.73.185:8080",
    "paths": {
        "/login": {
            "get": {
                "description": "用户登录请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户服务"
                ],
                "summary": "用户登录请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户结构体json",
                        "schema": {
                            "$ref": "#/definitions/people.User"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/login/register": {
            "get": {
                "description": "用户注册请求",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户服务"
                ],
                "summary": "用户注册请求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户id",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用户密码",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正确信息",
                        "schema": {
                            "$ref": "#/definitions/service.Response"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "people.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "wxOpenId": {
                    "type": "string"
                },
                "wxUnionId": {
                    "type": "string"
                }
            }
        },
        "service.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "相应码，错误为0，正确为1",
                    "type": "integer"
                },
                "msg": {
                    "description": "信息，错误为具体信息，正确为空",
                    "type": "string"
                }
            }
        }
    }
}